requirements.txt
Flask==2.0.1
numpy==1.21.2
tensorflow==2.6.0
Pillow==8.2.0
